app/requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
aniso8601==1.0.0
argparse==1.2.1
cffi==1.2.1
cryptography==1.0
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
itsdangerous==0.24
pyasn1==0.1.8
pycparser==2.14
pytz==2015.4
requests==2.7.0
six==1.9.0
wheel==0.24.0
wsgiref==0.1.2
